requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
pydantic==2.7.1
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.0
pyjwt==2.8.0
python-multipart==0.0.6
pydantic-settings==2.4.0
motor==3.3.1
pymongo==4.5.0
sentry-sdk ==1.40.6
bcrypt==4.0.1
jinja2==3.1.4